ensure that stdout of children are recombined at full lines
stdout is useless otherwise, because a line can be made of partial lines
from the child processes writing at same time as each other.

Results in a performance hit for 1 process, but comparable performance
for 4 or more processes.

Will likely add an option to disable this feature for maximum
performance, and also reimplement with non-blocking IO at some point.